Renamed Mana icon script to fit naming convention.
