+++ views/ManaView.php
